Feat: add  journal body
