+++ src/pages/users/project/[projectId]/journal/index.tsx
@@ -1,20 +1,15 @@
 import Head from 'next/head';
-import { FaArrowLeft } from 'react-icons/fa';
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations';
 import NavBar from '@/components/nav_bar/nav_bar';
 import { GetServerSideProps } from 'next';
 import ProjectSidebar from '@/components/project_sidebar/project_sidebar';
+import AddJournalBody from '@/components/add_journal_body/add_journal_body';
 
 interface IProjectJournalPageProps {
   projectId: string;
 }
 
 const ProjectJournalPage = ({ projectId }: IProjectJournalPageProps) => {
-  // ToDo: (20240621 - Julian) Replace with real data
-  const projectName = 'BAIFA';
-
-  const backClickHandler = () => window.history.back();
-
   return (
     <>
       <Head>
@@ -22,7 +17,7 @@ const ProjectJournalPage = ({ projectId }: IProjectJournalPageProps) => {
         <meta name="viewport" content="width=device-width, initial-scale=1" />
         <link rel="icon" href="/favicon/favicon.ico" />
         {/* TODO: (2024606 - Julian) i18n */}
-        <title>{projectName} Journal List - iSunFA</title>
+        <title>Project Add Journal - iSunFA</title>
       </Head>
 
       <div className="h-screen font-barlow">
@@ -31,33 +26,10 @@ const ProjectJournalPage = ({ projectId }: IProjectJournalPageProps) => {
         </div>
 
         <div className="flex w-full flex-1 flex-col overflow-x-hidden">
+          {/* Info: (20240621 - Julian) Sidebar */}
           <ProjectSidebar projectId={projectId} />
-          <div className="flex min-h-screen bg-gray-100">
-            <div className="my-120px flex-1 md:ml-80px">
-              <div className="flex flex-col px-16px md:px-60px">
-                <div className="flex w-full items-center justify-between">
-                  {/* Info: (2024621 - Julian) Title */}
-                  <div className="flex items-center gap-24px">
-                    <button
-                      type="button"
-                      onClick={backClickHandler}
-                      className="rounded border border-navyBlue p-12px text-navyBlue hover:border-primaryYellow hover:text-primaryYellow"
-                    >
-                      <FaArrowLeft />
-                    </button>
-
-                    <h1 className="text-base font-semibold text-text-neutral-secondary md:text-4xl">
-                      {projectName} - Add New Journal
-                    </h1>
-                  </div>
-                  {/* Info: (20240621 - Julian) Add new contract button (desktop) */}
-                </div>
-                {/* Info: (20240621 - Julian) Divider */}
-                <hr className="my-24px border border-divider-stroke-lv-4" />
-                {/* Info: (2024621 - Julian) Content */}
-              </div>
-            </div>
-          </div>
+          {/* Info: (20240621- Julian) Main */}
+          <AddJournalBody />
         </div>
       </div>
     </>
