+++ src/pages/users/project/[projectId]/journal/journal-list.tsx
@@ -1,13 +1,15 @@
 import Head from 'next/head';
+import Link from 'next/link';
+import { GetServerSideProps } from 'next';
 import { FaArrowLeft } from 'react-icons/fa';
+import { FiPlus } from 'react-icons/fi';
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations';
 import NavBar from '@/components/nav_bar/nav_bar';
-import { GetServerSideProps } from 'next';
 import ProjectSidebar from '@/components/project_sidebar/project_sidebar';
-import { FiPlus } from 'react-icons/fi';
 import { Button } from '@/components/button/button';
 import JournalListBody from '@/components/journal_list_body/journal_list_body';
 import { IDummyJournal } from '@/interfaces/journal';
+import { ISUNFA_ROUTE } from '@/constants/url';
 
 interface IProjectJournalListPageProps {
   projectId: string;
@@ -29,7 +31,7 @@ const ProjectJournalListPage = ({ projectId }: IProjectJournalListPageProps) => 
         <meta charSet="UTF-8" />
         <meta name="viewport" content="width=device-width, initial-scale=1" />
         <link rel="icon" href="/favicon/favicon.ico" />
-        {/* TODO: (2024606 - Julian) i18n */}
+        {/* TODO: (20240621 - Julian) i18n */}
         <title>{projectName} Journal List - iSunFA</title>
       </Head>
 
@@ -60,14 +62,16 @@ const ProjectJournalListPage = ({ projectId }: IProjectJournalListPageProps) => 
                   </div>
                   {/* Info: (20240621 - Julian) Add new contract button (desktop) */}
 
-                  <Button
-                    type="button"
-                    variant="tertiary"
-                    className="hidden items-center gap-4px px-4 py-8px md:flex"
-                  >
-                    <FiPlus size={24} />
-                    Add new journal
-                  </Button>
+                  <Link href={`${ISUNFA_ROUTE.PROJECT_LIST}/${projectId}/journal`}>
+                    <Button
+                      type="button"
+                      variant="tertiary"
+                      className="hidden items-center gap-4px px-4 py-8px md:flex"
+                    >
+                      <FiPlus size={24} />
+                      Add new journal
+                    </Button>
+                  </Link>
                   {/* Info: (20240621 - Julian) Add new contract button (mobile) */}
                   <Button
                     type="button"
